+++ world-chain-builder/crates/world/payload/src/lib.rs
@@ -1,3 +1,2 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 pub mod builder;
-pub mod stamp;
